standing up a prototype of matrix nio
