+++ Dockerfile.matrix
@@ -19,7 +19,7 @@ WORKDIR /app
 
 COPY . .
 
-RUN python3 -m pip install "matrix-nio[e2e]" &&\
+RUN python3 -m pip install -r matrix_requirements.txt --no-cache-dir &&\
     python3 -m pip install -e . --no-cache-dir
 
 RUN mkdir /home/nonroot && chown -R nonroot /home/nonroot /app
